requirements.txt
Flask==2.3.2
matplotlib==3.8.0
nltk==3.8.1
Pillow==10.0.1
mysql-connector-python~=9.1.0
PyPDF2~=3.0.1
flask
fitz
transformers
youtube-transcript-api
matplotlib
